requirements.txt
Flask==2.0.1
pandas==1.3.3
openpyxl==3.0.9
PyYAML==5.4.1
Werkzeug==2.0.1
xlsxwriter==3.0.2